services/scraper-service/requirements.txt
# Core dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# HTTP and web scraping
requests>=2.31.0
beautifulsoup4>=4.12.2
lxml>=4.9.3
selenium>=4.15.2
scrapy>=2.11.0
cloudscraper>=1.2.71

# Data processing
pandas>=2.2.0
numpy>=1.24.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.12.0

# Caching and messaging
redis>=4.5.0
celery>=5.3.0

# Monitoring and logging
prometheus-client>=0.17.0
loguru>=0.7.0

# Utilities
python-dotenv>=1.0.0
click>=8.1.0
